So, there was a bug in find linear part due to a recent change that re-ordered the scanorder from: edge, to two to any; to: edge, to one, to any.
To fix this bug I've reordered the scanorder to Edge->two->one.

The problem is now and then we'd start at a 2 neighbor where one branch
was not getting scanned becasue of the scan didnt reach around. this
occurs on inside edges.

Now, if branch chaser is on its outer loop, it will scan the masterpoint
and run the other way, then return those values, reverse them, and
prepend them.